% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSizeCRT.R
\name{checkSuff}
\alias{checkSuff}
\title{Check whether number of clusters sufficient to achieve specified Type
I error and power}
\usage{
checkSuff(
  outcome_type,
  m,
  max_i,
  sigma_vec = c(1, 1),
  p_vec = c(0, 0),
  rho,
  alpha = 0.05,
  beta = 0.1
)
}
\arguments{
\item{outcome_type}{\code{1=}continuous difference of means
\cr \code{2=}binary difference of proportions \cr \code{3=}binary odds ratio}

\item{m}{Number of clusters per arm.}

\item{sigma_vec}{Standard deviations for groups 1 and 2 (continuous case).}

\item{p_vec}{Probability of success for groups 1 and 2.}

\item{rho}{Intraclass correlation coefficient.}

\item{alpha}{Type I error, always one-sided. Default value is 0.05.}

\item{beta}{Type II error, default value is 0.1 (90\% power).}

\item{test_type}{\code{1=}one-sided test with early stopping for efficacy
\cr \code{2=}two-sided symmetric with early stopping for efficacy
\cr \code{3=}one-sided with early stopping for efficacy or futility
\cr \code{4=}two-sided with early stopping for efficacy or futility}

\item{delta}{Effect size for theta under alternative hypothesis.}
}
\value{
Binary variable corresponding to whether m clusters is sufficient to
achieve Type I error and power to detect an effect size of delta
}
\description{
Check whether number of clusters sufficient to achieve specified Type
I error and power
}
