% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSimCRT.R
\name{genBinCRT}
\alias{genBinCRT}
\title{Simulate cluster-randomized trial data with binary outcomes}
\usage{
genBinCRT(m = 1, n = 1, p_vec = c(0.5, 0.5), rho = NULL)
}
\arguments{
\item{m}{Number of clusters.}

\item{n}{Mean size of each cluster.}

\item{p_vec}{Probabilities of event for control and treatment groups.}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}
}
\value{
Simulated binary outcomes represented as an n x 4 matrix with the
 columns encoding the treatment arm, cluster, individual, and response.
}
\description{
Simulate cluster-randomized trial data with binary outcomes
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
