% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSimCRT.R
\name{genClusterSizes}
\alias{genClusterSizes}
\title{Generate varying cluster sizes for cluster-randomized trials}
\usage{
genClusterSizes(m, n, n_cv, n_min, n_max)
}
\arguments{
\item{m}{Number of clusters.}

\item{n}{Mean size of each cluster.}

\item{n_cv}{Coefficient of variation of cluster sizes.}

\item{n_min}{Minimum allowable cluster size.}

\item{n_max}{Maximum allowable cluster size.}
}
\value{
A vector of cluster sizes.
}
\description{
Generate varying cluster sizes for cluster-randomized trials
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
