% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSimCRT.R
\name{genContCRT}
\alias{genContCRT}
\title{Simulate cluster-randomized trial data with continuous outcomes}
\usage{
genContCRT(
  m = c(1, 1),
  m_alloc = c(0.5, 0.5),
  n = c(1, 1),
  n_cv = c(0, 0),
  mu_vec = c(0, 1),
  sigma_vec = c(1, 1),
  rho = c(0, 0)
)
}
\arguments{
\item{m}{Number of clusters.}

\item{m_alloc}{Allocation ratio of clusters per arm. Default is
\code{c(0.5, 0.5)}.}

\item{n}{Mean size of each cluster.}

\item{mu_vec}{Vector of means for control and treatment groups, respectively.}

\item{sigma_vec}{Vector of standard deviations for control and treatment
groups, respectively.}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}
}
\value{
Simulated continuous outcomes represented as an n x 4 matrix with the
 columns encoding the treatment arm, cluster, individual, and response.
}
\description{
Simulate cluster-randomized trial data with continuous outcomes
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
