% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSimBinCRT.R
\name{gsBinCRT}
\alias{gsBinCRT}
\alias{genBinCRT}
\title{Simulate cluster-randomized trial data with continuous outcomes}
\usage{
genBinCRT(
  m = 1,
  n = 1,
  p_vec = c(0.5, 0.5),
  rho = NULL,
  var_b = NULL,
  var_w = NULL
)
}
\arguments{
\item{m}{Number of clusters.}

\item{n}{Mean size of each cluster.}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}

\item{var_b}{Between-class variance.}

\item{var_w}{Within-class variance.}

\item{mu_vec}{Vector of means for groups 1 and 2, respectively.}

\item{sigma_vec}{Vector of standard deviations for groups 1 and 2,
respectively.}

\item{c}{Constant to convert probit to expit}
}
\value{
Results
}
\description{
Simulate cluster-randomized trial data with continuous outcomes
}
