% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsBounds1}
\alias{gsBounds1}
\title{Boundaries for 1-sided test with efficacy and futility stopping}
\usage{
gsBounds1(
  theta,
  I,
  falseneg,
  falsepos,
  binding = TRUE,
  tol = 1e-06,
  r = 42,
  printerr = 0
)
}
\description{
Boundaries for 1-sided test with efficacy and futility stopping
}
