% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsBounds2}
\alias{gsBounds2}
\title{Boundaries for 2-sided test with efficacy and futility stopping}
\usage{
gsBounds2(
  theta,
  I,
  falseneg,
  falsepos,
  binding = TRUE,
  tol = 1e-06,
  r = 42,
  printerr = 0
)
}
\description{
Boundaries for 2-sided test with efficacy and futility stopping
}
