% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSimContCRT.R
\name{gsContCRT}
\alias{gsContCRT}
\alias{genContCRT}
\title{Simulate cluster-randomized trial data with continuous outcomes}
\usage{
genContCRT(m = 1, n = 1, mu_vec = c(0, 1), sigma_vec = c(1, 1), rho = 0)
}
\arguments{
\item{m}{Number of clusters.}

\item{n}{Mean size of each cluster.}

\item{mu_vec}{Vector of means for groups 1 and 2, respectively.}

\item{sigma_vec}{Vector of standard deviations for groups 1 and 2,
respectively.}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}

\item{var_b}{Between-class variance.}

\item{var_w}{Within-class variance.}
}
\value{
Results
}
\description{
Simulate cluster-randomized trial data with continuous outcomes
}
