% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT-package.R
\docType{package}
\name{gsDesignCRT-package}
\alias{gsDesignCRT-package}
\alias{_PACKAGE}
\title{gsDesignCRT: Group Sequential Design for Cluster Randomized Trials}
\description{
Derives group sequential cluster randomized trial designs with continuous or binary outcomes. Based on the methods described in Jennison, Christopher and Turnbull, Bruce W., 2000, "Group Sequential Methods with Applications to Clinical Trials" and implemented in the package gsDesign (Anderson, 2024).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leejtding.github.io/gsDesignCRT/}
}

}
\author{
\strong{Maintainer}: Lee Ding \email{lee_ding@g.harvard.edu}

}
\keyword{internal}
