% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsDesignCRT}
\alias{gsDesignCRT}
\title{Compute stopping boundaries, maximum sample size, and expected sample
 sizes for a group sequential cluster randomized trial.}
\usage{
gsDesignCRT(
  k = 3,
  outcome_type = 1,
  test_type = 1,
  test_sides = 1,
  size_type = 1,
  recruit_type = 1,
  timing_type = 2,
  delta = 1,
  sigma_vec = c(1, 1),
  p_vec = c(0.5, 0.5),
  rho = 0,
  alpha = 0.05,
  beta = 0.1,
  m_fix = 1,
  n_fix = 1,
  info_timing = 1,
  size_timing = 1,
  alpha_sf = sfLDOF,
  alpha_sfpar = -4,
  beta_sf = sfLDOF,
  beta_sfpar = -4,
  tol = 1e-06,
  r = 18
)
}
\arguments{
\item{k}{Number of analyses planned, including interim and final.}

\item{outcome_type}{\code{1=}continuous difference of means
\cr \code{2=}binary difference of proportions}

\item{test_type}{\code{1=} early stopping for efficacy only
\cr \code{2=} early stopping for binding futility only
\cr \code{3=} early stopping for non-binding futility only
\cr \code{4=} early stopping for either efficacy or binding futility
\cr \code{5=} early stopping for either efficacy or non-binding futility}

\item{test_sides}{\code{1=} one-sided test \cr \code{2=} two-sided test}

\item{size_type}{\code{1=}clusters per arm \cr \code{2=}cluster size}

\item{recruit_type}{\code{1=}recruit clusters with fixed sizes
\cr \code{2=}recruit individuals into fixed number of clusters
\cr \code{3=}recruit both clusters and individuals}

\item{timing_type}{\code{1=} maximum and expected sample sizes based on
specified information levels in \code{info_timing} \cr \code{2=} maximum
sample size based on specified information levels in \code{info_timing} and
expected sample sizes based on specified sample size levels in
\code{size_timing}}

\item{delta}{Effect size for theta under alternative hypothesis. Must be > 0.}

\item{sigma_vec}{Standard deviations for control and treatment groups
(continuous case).}

\item{p_vec}{Probabilities of event for control and treatment groups
(binary case).}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}

\item{alpha}{Type I error, default value is 0.05.}

\item{beta}{Type II error, default value is 0.1 (90\% power).}

\item{m_fix}{Number of clusters; used to find maximum size of each cluster.}

\item{n_fix}{Mean size of each cluster; used to find maximum number of
clusters per arm.}

\item{info_timing}{Sets timing of interim analyses based on information
levels. Default of 1 produces analyses at equal-spaced increments.
Otherwise, this is a vector of length \code{k} or \code{k-1}. The values
should satisfy \code{0 < info_timing[1] < info_timing[2] < ... <
info_timing[k-1] < info_timing[k]=1}.}

\item{size_timing}{Sets timing of interim analyses based on sample size
levels if \code{timing_type = 2}. Default of 1 produces analyses at
equal-spaced increments. Otherwise, this is a vector of length \code{k} or
\code{k-1}. The values should satisfy \code{0 < size_timing[1] <
size_timing[2] < ... < size_timing[k-1] < size_timing[k]=1}.}

\item{alpha_sf}{A spending function or a character string indicating an
upper boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds,
\dQuote{OF} for O'Brien-Fleming bounds, and \dQuote{Pocock} for Pocock
bounds). The default value is \code{sfLDOF} which is a Lan-DeMets
O'Brien-Fleming spending function. See details,
\code{vignette("SpendingFunctionOverview")}, manual and examples.}

\item{alpha_sfpar}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with a
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions. The parameter \code{alpha_sfpar} specifies any parameters needed
for the spending function specified by \code{alpha_sf}; this will be ignored
for spending functions (\code{sfLDOF}, \code{sfLDPocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{beta_sf}{A spending function or a character string indicating an
lower boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds,
\dQuote{OF} for O'Brien-Fleming bounds, and \dQuote{Pocock} for Pocock
bounds). The default value is \code{sfLDOF} which is a Lan-DeMets
O'Brien-Fleming spending function. See details,
\code{vignette("SpendingFunctionOverview")}, manual and examples.}

\item{beta_sfpar}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with a
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions. The parameter \code{beta_sfpar} specifies any parameters needed
for the spending function specified by \code{beta_sf}; this will be ignored
for spending functions (\code{sfLDOF}, \code{sfLDPocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{tol}{Tolerance for error (default is 0.000001). Normally this will not
be changed by the user.  This does not translate directly to number of
digits of accuracy, so use extra decimal places.}

\item{r}{Integer value controlling grid for numerical integration as in
Jennison and Turnbull (2000); default is 18, range is 1 to 80.  Larger
values provide larger number of grid points and greater accuracy.  Normally
\code{r} will not be changed by the user.}
}
\value{
Object containing the following elements: \item{k}{As
 input.} \item{outcome_type}{As input.} \item{test_type}{As input.}
 \item{test_sides}{As input.} \item{size_type}{As input.}
 \item{recruit_type}{As input.} \item{timing_type}{As input.} \item{delta}{As
 input.} \item{sigma_vec}{As input.} \item{p_vec}{As input.} \item{rho}{As
 input.} \item{alpha}{As input.} \item{beta}{As input.} \item{info_timing}{As
 input.} \item{size_timing}{As input.} \item{i}{Fisher information at each
 planned interim analysis based on \code{timing_type}.} \item{max_i}{Maximum
 information corresponding to design specifications.} \item{m}{Number of
 clusters per arm at each planned interim analysis.} \item{max_m}{Maximum
 number of clusters per arm.} \item{e_m}{A vector of length 2 with expected
 number of clusters per arm under the null and alternative hypotheses. For
 simplicity, the expected sizes with non-binding futility boundaries are
 calculated assuming the boundaries are binding futility.} \item{n}{Average
 cluster size at each planned interim analysis.} \item{max_n}{Maximum cluster
 size.} \item{e_m}{A vector of length 2 with expected cluster sizes under the
 null and alternative hypotheses. For simplicity, the expected sizes with
 non-binding futility boundaries are calculated assuming the futility
 boundaries are binding.} \item{max_total}{Maximum number of individuals in
 the trial.} \item{e_total}{A vector of length 2 with expected number of
 individuals in the trial under the null and alternative hypotheses. For
 simplicity, the expected sizes with non-binding futility boundaries are
 calculated assuming the futility boundaries are binding.} \item{sufficient}{
 Value denoting whether calculated sample size will be sufficient to achieve
 specified Type I error rate and power given the trial specifications.}
 \item{lower_bound}{Calculated lower futility boundaries under analysis
 schedule specified by \code{timing_type}} \item{upper_bound}{Calculated
 upper efficacy boundaries under analysis schedule specified by
 \code{timing_types}.} \item{tol}{As input.} \item{r}{As input.}
}
\description{
\code{gsDesignCRT()} is used to determine the maximum sample
 size needed for a specified parallel group sequential cluster randomized
 trial to detect a clinically meaningful effect size with some Type I error
 rate and power. Code adapted from gsDesign package.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
 Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
