% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsPlotCRT}
\alias{gsPlotCRT}
\title{Plot group sequential CRT design}
\usage{
gsPlotCRT(x)
}
\arguments{
\item{x}{Object of class \code{gsDesignCRT} created by \code{gsDesignCRT()}.}
}
\value{
Plot of stopping boundaries for the group sequential CRT design
}
\description{
Plot group sequential CRT design
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
