% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsPrintCRT}
\alias{gsPrintCRT}
\title{Print group sequential CRT design}
\usage{
gsPrintCRT(x)
}
\arguments{
\item{x}{Object of class \code{gsDesignCRT} created by \code{gsDesignCRT()}.}
}
\value{
Printed output of design parameters and table of stopping boundaries
}
\description{
Print group sequential CRT design
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
