% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsProbability1}
\alias{gsProbability1}
\title{Boundary Crossing Probabilities for 1-sided test}
\usage{
gsProbability1(k = 0, theta, I, a, b, r = 42, d = NULL, overrun = 0)
}
\description{
Boundary Crossing Probabilities for 1-sided test
}
