% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsProbability2}
\alias{gsProbability2}
\title{Boundary Crossing Probabilities for 2-sided test}
\usage{
gsProbability2(k = 0, theta, I, a, b, r = 42, d = NULL, overrun = 0)
}
\description{
Boundary Crossing Probabilities for 2-sided test
}
