% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSimBinCRT.R
\name{gsSimBinCRT}
\alias{gsSimBinCRT}
\title{Simulate group sequential cluster-randomized trial with continuous
outcomes}
\usage{
gsSimBinCRT(
  t,
  data,
  test_type,
  recruit_type,
  icc_type,
  stat_type,
  delta,
  p_vec,
  rho,
  alpha = 0.05,
  beta = 0.1,
  m,
  n,
  ifrac_size_m = NULL,
  ifrac_size_n = NULL,
  alpha_sf,
  alpha_sfpar = -4,
  beta_sf,
  beta_sfpar = -4,
  tol = 1e-06,
  r = 42
)
}
\arguments{
\item{t}{Number of analyses planned, including interim and final.}

\item{data}{Simulated trial data.}

\item{test_type}{\code{1=}one-sided test with early stopping for efficacy
\cr \code{2=}two-sided symmetric with early stopping for efficacy
\cr \code{3=}one-sided with early stopping for efficacy or futility
\cr \code{4=}two-sided with early stopping for efficacy or futility}

\item{recruit_type}{\code{1=}clusters \cr \code{2=}individuals per cluster}

\item{alpha}{Type I error}

\item{beta}{Type II error}

\item{m}{Number of clusters.}

\item{n}{Mean size of each cluster.}

\item{alpha_sf}{A spending function or a character string indicating a
boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds, \dQuote{OF} for
O'Brien-Fleming bounds and \dQuote{Pocock} for Pocock bounds).  For
one-sided and symmetric two-sided testing is used to completely specify
spending (\code{test.type=1, 2}), \code{sfu}.  The default value is
\code{sfHSD} which is a Hwang-Shih-DeCani spending function.  See details,
\code{vignette("SpendingFunctionOverview")}, manual and examples.}

\item{alpha_sfpar}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with the default
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions.  The parameter \code{sfupar} specifies any parameters needed for
the spending function specified by \code{sfu}; this will be ignored for
spending functions (\code{sfLDOF}, \code{sfLDPocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{beta_sf}{Specifies the spending function for lower boundary crossing
probabilities when asymmetric, two-sided testing is performed
(\code{test.type = 3}, \code{4}, \code{5}, or \code{6}).  Unlike the upper
bound, only spending functions are used to specify the lower bound.  The
default value is \code{sfHSD} which is a Hwang-Shih-DeCani spending
function.  The parameter \code{sfl} is ignored for one-sided testing
(\code{test.type=1}) or symmetric 2-sided testing (\code{test.type=2}).  See
details, spending functions, manual and examples.}

\item{beta_sfpar}{Real value, default is \eqn{-2}, which, with the default
Hwang-Shih-DeCani spending function, specifies a less conservative spending
rate than the default for the upper bound.}

\item{tol}{Tolerance for error (default is 0.000001). Normally this will not
be changed by the user.  This does not translate directly to number of
digits of accuracy, so use extra decimal places.}

\item{r}{Integer value controlling grid for numerical integration as in
Jennison and Turnbull (2000); default is 18, range is 1 to 80.  Larger
values provide larger number of grid points and greater accuracy.  Normally
\code{r} will not be changed by the user.}

\item{ifrac_size}{Sample size at each interim look}
}
\value{
Results
}
\description{
Simulate group sequential cluster-randomized trial with continuous
outcomes
}
