% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSimCRT.R
\name{gsSimContCRT}
\alias{gsSimContCRT}
\title{Simulate group sequential cluster-randomized trial with continuous
outcomes.}
\usage{
gsSimContCRT(
  k,
  data,
  test_type,
  test_sides,
  recruit_type,
  stat_type,
  balance_size,
  precompute = FALSE,
  delta,
  sigma_vec = c(1, 1),
  rho,
  alpha = 0.05,
  beta = 0.1,
  lower_bound = NULL,
  upper_bound = NULL,
  m_max = 1,
  n_max = 1,
  schedule_m = NULL,
  schedule_n = NULL,
  alpha_sf,
  alpha_sfpar = -4,
  beta_sf,
  beta_sfpar = -4,
  tol = 1e-06,
  r = 18
)
}
\arguments{
\item{k}{Number of analyses planned, including interim and final.}

\item{data}{Simulated continuous outcomes. Should be an n x 4 matrix with the
columns encoding the treatment arm, cluster, individual, and response.}

\item{test_type}{\code{1=} early stopping for efficacy only
\cr \code{2=} early stopping for binding futility only
\cr \code{3=} early stopping for non-binding futility only
\cr \code{4=} early stopping for either efficacy or binding futility
\cr \code{5=} early stopping for either efficacy or non-binding futility}

\item{test_sides}{\code{1=} one-sided test \cr \code{2=} two-sided test}

\item{recruit_type}{\code{1=}by clusters with all individuals recruited
\cr \code{2=}by individuals in recruited cluster
\cr \code{3=}by both clusters and individuals in clusters}

\item{stat_type}{\code{1=} Z-test with known variance and ICC
\cr \code{2=} Z-test with re-estimated variance and ICC
\cr \code{3=} t-test with re-estimated variance and ICC}

\item{balance_size}{\code{1=} exact sample increments according to the
scheduled interim analyses.
\cr \code{2=} randomized sample increments from multinomial distribution
according to the scheduled interim analyses.}

\item{precompute}{Use pre-computed stopping boundaries if true.}

\item{delta}{Effect size for theta under alternative hypothesis.}

\item{sigma_vec}{Standard deviations for control and treatment groups.}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}

\item{alpha}{Desired Type I error, always one-sided. Default value is 0.05.}

\item{beta}{Desired Type II error, default value is 0.1 (90\% power).}

\item{lower_bound}{Pre-computed lower futility boundaries at the specified
interim analyses. Must be specified if precompute is TRUE. NULL otherwise.}

\item{upper_bound}{Pre-computed upper efficacy boundaries at the specified
interim analyses. Must be specified if precompute is TRUE. NULL otherwise.}

\item{m_max}{Number of clusters.}

\item{n_max}{Mean size of each cluster.}

\item{schedule_m}{Number of clusters at each interim look. Interim analyses
will be conducted according to the information levels in \code{schedule_m}
and \code{schedule_n} if provided. Otherwise, interim analyses will be
conducted at equal-sized sample increments according to \code{recruit_type}.}

\item{schedule_n}{Average cluster size at each interim look. Interim
analyses will be conducted according to the information levels in
\code{schedule_m} and \code{schedule_n} if provided. Otherwise, interim
analyses will be conducted at equal-sized sample increments according to
\code{recruit_type}.}

\item{alpha_sf}{A spending function or a character string indicating an
upper boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds,
\dQuote{OF} for O'Brien-Fleming bounds, and \dQuote{Pocock} for Pocock
bounds). The default value is \code{sfLDOF} which is a Lan-DeMets
O'Brien-Fleming spending function. See details,
\code{vignette("SpendingFunctionOverview")}, manual and examples.}

\item{alpha_sfpar}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with a
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions. The parameter \code{alpha_sfpar} specifies any parameters needed
for the spending function specified by \code{alpha_sf}; this will be ignored
for spending functions (\code{sfLDOF}, \code{sfLDPocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{beta_sf}{A spending function or a character string indicating an
lower boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds,
\dQuote{OF} for O'Brien-Fleming bounds, and \dQuote{Pocock} for Pocock
bounds). The default value is \code{sfLDOF} which is a Lan-DeMets
O'Brien-Fleming spending function. See details,
\code{vignette("SpendingFunctionOverview")}, manual and examples.}

\item{beta_sfpar}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with a
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions. The parameter \code{beta_sfpar} specifies any parameters needed
for the spending function specified by \code{beta_sf}; this will be ignored
for spending functions (\code{sfLDOF}, \code{sfLDPocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{tol}{Tolerance for error (default is 0.000001). Normally this will not
be changed by the user.  This does not translate directly to number of
digits of accuracy, so use extra decimal places.}

\item{r}{Integer value controlling grid for numerical integration as in
Jennison and Turnbull (2000); default is 18, range is 1 to 80.  Larger
values provide larger number of grid points and greater accuracy.  Normally
\code{r} will not be changed by the user.}
}
\value{
Object containing the following elements: \item{reject}{Whether the
null hypothesis was rejected in the simulated trial.} \item{k_i}{Interim
analysis at which simulated trial was stopped.} \item{m_i}{Number of clusters
per arm when the simulated trial was stopped.} \item{n_i}{Average number of
individuals per cluster when the simulated trial was stopped.}
\item{total_i}{Total number of individuals per arm when the simulated trial
was stopped.} \item{i_frac}{Information fraction when the simulated trial
was stopped.}
}
\description{
Simulate group sequential cluster-randomized trial with continuous
outcomes.
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
