% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSizeCRT.R
\name{gsSizeCRT}
\alias{gsSizeCRT}
\title{Compute sample size for group sequential cluster-randomized trial}
\usage{
gsSizeCRT(
  t = 3,
  test_type = 1,
  outcome_type = 1,
  size_type = 1,
  stat_type = 1,
  delta = 1,
  sigma_vec = NULL,
  p_vec = NULL,
  rho = 0,
  alpha = 0.05,
  beta = 0.1,
  m_fix = 1,
  n_fix = 1,
  timing = 1,
  alpha_sf = sfLDOF,
  alpha_sfpar = -4,
  beta_sf = sfLDOF,
  beta_sfpar = -4,
  tol = 1e-06,
  r = 42
)
}
\arguments{
\item{t}{Number of analyses planned, including interim and final.}

\item{test_type}{\code{1=}one-sided test with early stopping for efficacy
\cr \code{2=}two-sided test with early stopping for efficacy
\cr \code{3=}one-sided test with efficacy or binding futility stopping
\cr \code{4=}two-sided test with efficacy or binding futility stopping
\cr \code{5=}one-sided test with efficacy or non-binding futility stopping
\cr \code{6=}two-sided test with efficacy or non-binding futility stopping}

\item{outcome_type}{\code{1=}continuous difference of means
\cr \code{2=}binary difference of proportions \cr \code{3=}binary odds ratio}

\item{size_type}{\code{1=}clusters \cr \code{2=}individuals per cluster}

\item{delta}{Effect size for theta under alternative hypothesis.}

\item{sigma_vec}{Standard deviations for groups 1 and 2 (continuous case)}

\item{p_vec}{Probabilities of event for groups 1 and 2 (binary case)}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}

\item{alpha}{Type I error, always one-sided. Default value is 0.05.}

\item{beta}{Type II error, default value is 0.1 (90\% power).}

\item{m_fix}{Number of clusters; used to find maximum group sequential
size of each cluster}

\item{n_fix}{Mean size of each cluster; used to find maximum group sequential
number of clusters}

\item{timing}{Sets relative timing of interim analyses. Default of 1
produces equally spaced analyses.  Otherwise, this is a vector of length
\code{t} or \code{t-1}.  The values should satisfy \code{0 < timing[1] <
timing[2] < ... < timing[t-1] < timing[t]=1}.}

\item{alpha_sf}{A spending function or a character string indicating a
boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds, \dQuote{OF} for
O'Brien-Fleming bounds and \dQuote{Pocock} for Pocock bounds).  For
one-sided and symmetric two-sided testing is used to completely specify
spending (\code{test.type=1, 2}), \code{sfu}.  The default value is
\code{sfHSD} which is a Hwang-Shih-DeCani spending function.  See details,
\code{vignette("SpendingFunctionOverview")}, manual and examples.}

\item{alpha_sfpar}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with the default
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions.  The parameter \code{sfupar} specifies any parameters needed for
the spending function specified by \code{sfu}; this will be ignored for
spending functions (\code{sfLDOF}, \code{sfLDPocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{beta_sf}{Specifies the spending function for lower boundary crossing
probabilities when asymmetric, two-sided testing is performed
(\code{test.type = 3}, \code{4}, \code{5}, or \code{6}).  Unlike the upper
bound, only spending functions are used to specify the lower bound.  The
default value is \code{sfHSD} which is a Hwang-Shih-DeCani spending
function.  The parameter \code{sfl} is ignored for one-sided testing
(\code{test.type=1}) or symmetric 2-sided testing (\code{test.type=2}).  See
details, spending functions, manual and examples.}

\item{beta_sfpar}{Real value, default is \eqn{-2}, which, with the default
Hwang-Shih-DeCani spending function, specifies a less conservative spending
rate than the default for the upper bound.}

\item{tol}{Tolerance for error (default is 0.000001). Normally this will not
be changed by the user.  This does not translate directly to number of
digits of accuracy, so use extra decimal places.}

\item{r}{Integer value controlling grid for numerical integration as in
Jennison and Turnbull (2000); default is 18, range is 1 to 80.  Larger
values provide larger number of grid points and greater accuracy.  Normally
\code{r} will not be changed by the user.}
}
\value{
Object with maximum information sample size and expected sample size
under null and alternate hypotheses
}
\description{
\code{gsSizeCRT()} is used to trial size required for a
cluster-randomized group sequential design.
}
