% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{gsUpper2}
\alias{gsUpper2}
\title{Boundary derivation - upper boundary}
\usage{
gsUpper2(theta, I, a, falsepos, tol = 1e-06, r = 42, printerr = 0)
}
\description{
Boundary derivation - upper boundary
}
