% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSizeCRT.R
\name{mContDiff}
\alias{mContDiff}
\title{Convert information to number of clusters for difference of continuous
outcomes}
\usage{
mContDiff(i, n, sigma_vec, rho)
}
\arguments{
\item{i}{Fisher information.}

\item{n}{Mean cluster size.}

\item{sigma_vec}{Standard deviations for groups 1 and 2 (continuous case).}

\item{rho}{Intraclass correlation coefficient.}
}
\value{
Number of clusters corresponding to inputs.
}
\description{
Convert information to number of clusters for difference of continuous
outcomes
}
