% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSizeCRT.R
\name{mPropDiff}
\alias{mPropDiff}
\title{Convert information to number of clusters for difference of
proportions}
\usage{
mPropDiff(i, n, p_vec, rho)
}
\arguments{
\item{i}{Fisher information.}

\item{n}{Mean cluster size.}

\item{p_vec}{Probability of success for groups 1 and 2.}

\item{rho}{Intraclass correlation coefficient.}
}
\value{
Number of clusters corresponding to inputs.
}
\description{
Convert information to number of clusters for difference of
proportions
}
