% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSizeCRT.R
\name{nContDiff}
\alias{nContDiff}
\title{Convert information to average cluster size for difference of
continuous outcomes}
\usage{
nContDiff(i, m, sigma_vec, rho)
}
\arguments{
\item{i}{Fisher information.}

\item{m}{Number of clusters per arm.}

\item{sigma_vec}{Standard deviations for groups 1 and 2 (continuous case).}

\item{rho}{Intraclass correlation coefficient.}
}
\value{
Average cluster size corresponding to inputs.
}
\description{
Convert information to average cluster size for difference of
continuous outcomes
}
