% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSizeCRT.R
\name{nPropDiff}
\alias{nPropDiff}
\title{Convert information to average cluster size for difference of
proportions}
\usage{
nPropDiff(i, m, p_vec, rho)
}
\arguments{
\item{i}{Fisher information.}

\item{m}{Number of clusters per arm.}

\item{p_vec}{Probability of success for groups 1 and 2.}

\item{rho}{Intraclass correlation coefficient.}
}
\value{
Average cluster size corresponding to inputs.
}
\description{
Convert information to average cluster size for difference of
proportions
}
