% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesignCRT.R
\name{print.gsDesignCRT}
\alias{print.gsDesignCRT}
\alias{plot.gsDesignCRT}
\title{Print group sequential CRT design}
\usage{
\method{print}{gsDesignCRT}(x)

\method{plot}{gsDesignCRT}(x)
}
\arguments{
\item{x}{Object of class \code{gsDesignCRT} created by \code{gsDesignCRT()}.}
}
\value{
Printed output of design parameters and table of stopping boundaries

Printed output of design parameters and table of stopping boundaries
}
\description{
Print method for \code{gsDesignCRT} objects.

Print method for \code{gsDesignCRT} objects.
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
