% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSpending.R
\name{sfPoints}
\alias{sfPoints}
\title{Pointwise Spending Function}
\usage{
sfPoints(alpha, t, param)
}
\arguments{
\item{alpha}{Real value \eqn{> 0} and no more than 1. Normally,
\code{alpha=0.025} for one-sided Type I error specification or
\code{alpha=0.1} for Type II error specification. However, this could be set
to 1 if for descriptive purposes you wish to see the proportion of spending
as a function of the proportion of sample size/information.}

\item{t}{A vector of points with increasing values from >0 and <=1.  Values
of the proportion of sample size/information for which the spending function
will be computed.}

\item{param}{A vector of the same length as \code{t} specifying the
cumulative proportion of spending to corresponding to each point in
\code{t}; must be >=0 and <=1.}
}
\value{
An object of type \code{spendfn}. See
\code{vignette("SpendingFunctionOverview")} for further details.
}
\description{
The function \code{sfPoints} implements a spending function with
values specified for an arbitrary set of specified points. It is now
recommended to use sfLinear rather than sfPoints. If using \code{sfPoints()}
in a design, it is recommended to specify how to interpolate between the
specified points (e.g,, linear interpolation); also consider fitting smooth
spending functions; see \code{vignette("SpendingFunctionOverview")}.
}
\note{
The gsDesign technical manual is available at
  \url{https://keaven.github.io/gsd-tech-manual/}.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\seealso{
\code{vignette("SpendingFunctionOverview")},
\code{\link{gsDesignCRT}}, \code{vignette("gsDesignCRTPackageOverview")},
\link{sfLogistic}
}
\author{
Keaven Anderson \email{keaven_anderson@merck.com}
}
\keyword{design}
