% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSizeCRT.R
\name{sizeInfoFrac}
\alias{sizeInfoFrac}
\title{Convert information fraction to sample size}
\usage{
sizeInfoFrac(size_type, ifrac, m_max = 1, n_max = 1, rho)
}
\arguments{
\item{size_type}{\code{1=}clusters \cr \code{2=}individuals per cluster}

\item{ifrac}{information fraction I_{t} / I_{max}}

\item{rho}{Intraclass correlation coefficient.}

\item{n}{Mean cluster size.}

\item{p_vec}{Probability of success for groups 1 and 2.}
}
\value{
Sample size corresponding to inputs.
}
\description{
Convert information fraction to sample size
}
