% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSpending.R
\name{spendingFunction}
\alias{spendingFunction}
\title{Spending Function}
\usage{
spendingFunction(alpha, t, param)
}
\arguments{
\item{alpha}{Real value \eqn{> 0} and no more than 1. Defaults in calls to
\code{gsDesignCRT()} are \code{alpha=0.025} for one-sided Type I error
specification and \code{alpha=0.1} for Type II error specification.
However, this could be set to 1 if, for descriptive purposes, you wish to
see the proportion of spending as a function of the proportion of sample
size/information.}

\item{t}{A vector of points with increasing values from 0 to 1, inclusive.
Values of the proportion of sample size/information for which the spending
function will be computed.}

\item{param}{A single real value or a vector of real values specifying the
spending function parameter(s); this must be appropriately matched to the
spending function specified.}
}
\value{
\code{spendingFunction} and spending functions in general produce an
object of type \code{spendfn}.
\item{name}{A character string with the name of the spending function.}
\item{param}{any parameters used for the spending function.}
\item{parname}{a character string or strings with the name(s) of
the parameter(s) in \code{param}.}
\item{sf}{the spending function specified.}
\item{spend}{a vector of cumulative spending values corresponding to
the input values in \code{t}.}
\item{bound}{this is null when returned from the spending function,
but is set in \code{gsDesignCRT()} if the spending function is called
from there.  Contains z-values for bounds of a design.}
\item{prob}{this is null when returned from the spending function,
but is set in \code{gsDesignCRT()} if the spending function is called
from there.  Contains probabilities of boundary crossing at \code{i}-th
analysis for \code{j}-th theta value input to \code{gsDesignCRT()} in
\code{prob[i,j]}.}
}
\description{
Spending Function
}
\note{
The gsDesign technical manual is available at
  \url{https://keaven.github.io/gsd-tech-manual/}.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\seealso{
\code{\link{gsDesignCRT}}, \code{\link{sfHSD}},
\code{\link{sfPower}}, \code{\link{sfLogistic}}, \code{\link{sfExponential}},
\code{\link{sfTruncated}}, \code{vignette("gsDesignCRTPackageOverview")}
}
\author{
Keaven Anderson \email{keaven_anderson@merck.com}
}
\keyword{design}
